refactor(ai-store): modularize state and action concerns
